requirements.txt
# Core
anthropic>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Optional but recommended
lxml>=5.0.0          # faster HTML parser for BeautifulSoup
